+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
 - Added github action to schedule daily and on demand rebooots
+
+### Changed
+- Update leaves using bulkWrite
+
 ## [5.0.3] - 2022-06-20
 ### Fixed
 - Return correct solana network name